encrypt credentials, IMDSv2 support

+++ src/main/perl/t/04-process.t
@@ -12,14 +12,19 @@ BEGIN {
   use_ok('Amazon::Credentials');
 }
 
-my $home = mkdtemp("amz-credentials-XXXXX");
+my $home = mkdtemp('amz-credentials-XXXXX');
 
 my $credentials_file = eval {
   mkdir "$home/.aws";
-  
-  open (my $fh, ">$home/.aws/credentials") or BAIL_OUT("could not create temporary credentials file");
+
+  open( my $fh, '>', "$home/.aws/credentials" )
+    or BAIL_OUT('could not create temporary credentials file');
+
   my $process = getcwd . '/get-creds-from-process';
-  
+
+  BAIL_OUT('cannot execute $process')
+    if !-x $process;
+
   print $fh <<eot;
 [profile foo]
 credential_process = $process
@@ -30,7 +35,7 @@ eot
   "$home/.aws/credentials";
 };
 
-$ENV{HOME} = "$home";
+$ENV{HOME}        = $home;
 $ENV{AWS_PROFILE} = undef;
 
 my $creds = Amazon::Credentials->new(
@@ -42,13 +47,9 @@ my $creds = Amazon::Credentials->new(
 
 ok( ref($creds), 'find credentials' );
 
-is($creds->get_aws_access_key_id, 'aws-access-key-id', 'get from process');
-is($creds->get_region, 'us-east-1', 'region');
-
+is( $creds->get_aws_access_key_id, 'aws-access-key-id', 'get from process' );
+is( $creds->get_region,            'us-east-1',         'region' );
 
 END {
-  eval {
-    rmtree($home)
-      if $home;
-  };
+  eval { rmtree($home) if $home; };
 }
